tests/unit/test_tasks/test_worker.py
"""
Unit tests for Celery worker configuration.
"""
import pytest
from unittest.mock import Mock, patch
from app.worker import celery_app


class TestCeleryWorkerConfiguration:
    """Test Celery worker configuration."""
    
    def test_celery_app_name(self):
        """Test Celery app has correct name."""
        assert celery_app.main == "kyc_worker"
    
    def test_celery_app_broker_configuration(self):
        """Test Celery broker configuration."""
        # The broker URL should be set from settings
        assert celery_app.conf.broker_url is not None
    
    def test_celery_app_backend_configuration(self):
        """Test Celery result backend configuration."""
        # The result backend should be set from settings
        assert celery_app.conf.result_backend is not None
    
    def test_task_serialization_settings(self):
        """Test task serialization settings."""
        assert celery_app.conf.task_serializer == "json"
        assert "json" in celery_app.conf.accept_content
        assert celery_app.conf.result_serializer == "json"
    
    def test_timezone_settings(self):
        """Test timezone configuration."""
        assert celery_app.conf.timezone == "UTC"
        assert celery_app.conf.enable_utc is True
    
    def test_task_routing_configuration(self):
        """Test task routing is configured."""
        task_routes = celery_app.conf.task_routes
        
        assert "app.tasks.kyc_tasks.*" in task_routes
        assert task_routes["app.tasks.kyc_tasks.*"]["queue"] == "kyc_queue"
        
        assert "app.tasks.webhook_tasks.*" in task_routes
        assert task_routes["app.tasks.webhook_tasks.*"]["queue"] == "webhook_queue"
    
    def test_worker_settings(self):
        """Test worker configuration settings."""
        assert celery_app.conf.worker_prefetch_multiplier == 1
        assert celery_app.conf.worker_max_tasks_per_child == 1000
        assert celery_app.conf.worker_disable_rate_limits is False
    
    def test_task_execution_settings(self):
        """Test task execution settings."""
        assert celery_app.conf.task_acks_late is True
        assert celery_app.conf.task_reject_on_worker_lost is True
    
    def test_monitoring_settings(self):
        """Test monitoring configuration."""
        assert celery_app.conf.worker_send_task_events is True
        assert celery_app.conf.task_send_sent_event is True
    
    def test_result_expiration(self):
        """Test result expiration setting."""
        assert celery_app.conf.result_expires == 3600  # 1 hour
    
    def test_included_tasks(self):
        """Test that task modules are included."""
        includes = celery_app.conf.include
        
        assert "app.tasks.kyc_tasks" in includes
        assert "app.tasks.webhook_tasks" in includes
    
    def test_beat_schedule_initialization(self):
        """Test that beat schedule is initialized."""
        # Should be empty dict initially
        assert celery_app.conf.beat_schedule == {}


class TestCeleryTaskDiscovery:
    """Test Celery task auto-discovery."""
    
    @patch.object(celery_app, 'autodiscover_tasks')
    def test_autodiscover_tasks_called(self, mock_autodiscover):
        """Test that autodiscover_tasks is called."""
        # Import the module to trigger autodiscovery
        import app.worker
        
        # Verify autodiscover was called
        # Note: This test verifies the configuration, actual autodiscovery
        # happens when the module is imported
        assert hasattr(celery_app, 'autodiscover_tasks')


class TestCeleryAppIntegration:
    """Test Celery app integration with settings."""
    
    @patch('app.worker.settings')
    def test_broker_url_from_settings(self, mock_settings):
        """Test that broker URL comes from settings."""
        mock_settings.CELERY_BROKER_URL = "pyamqp://test@localhost:5672//"
        
        # Re-import to get updated settings
        import importlib
        import app.worker
        importlib.reload(app.worker)
        
        # The actual test would require more complex setup to verify
        # settings integration, so we just verify the structure exists
        assert hasattr(app.worker, 'celery_app')
    
    @patch('app.worker.settings')
    def test_result_backend_from_settings(self, mock_settings):
        """Test that result backend comes from settings."""
        mock_settings.CELERY_RESULT_BACKEND = "redis://test:6379/1"
        
        # Import the module to verify structure exists
        import app.worker
        assert hasattr(app.worker, 'celery_app')


class TestCeleryConfiguration:
    """Test specific Celery configuration values."""
    
    def test_result_backend_transport_options(self):
        """Test result backend transport options."""
        transport_options = celery_app.conf.result_backend_transport_options
        
        assert "visibility_timeout" in transport_options
        assert transport_options["visibility_timeout"] == 3600
    
    def test_task_routes_structure(self):
        """Test task routes have correct structure."""
        task_routes = celery_app.conf.task_routes
        
        # Verify it's a dictionary
        assert isinstance(task_routes, dict)
        
        # Verify each route has queue configuration
        for pattern, config in task_routes.items():
            assert "queue" in config
            assert isinstance(config["queue"], str)
    
    def test_serialization_configuration(self):
        """Test serialization configuration is secure."""
        # Only JSON should be accepted for security
        assert celery_app.conf.task_serializer == "json"
        assert celery_app.conf.result_serializer == "json"
        assert celery_app.conf.accept_content == ["json"]