requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Async processing
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0

# Authentication and Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-multipart>=0.0.6

# HTTP client
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0
rich>=13.7.0